+++ CHANGELOG.rst
@@ -18,6 +18,7 @@
 - remove `IForbidden`
 - by default, provide an async queue utility
 - move `guillotina.files.CloudFileField` to `guillotina.fields.CloudFileField`
+- fix deserialization with BucketListField
 
 
 3.3.12 (2018-05-30)